v3.0
ОНО КОМПИЛИРУЕТСЯ!!!1!!!11!!!!1
~~~
СВЕРКАЛИ ГОРЫ И ЛЕСА
СВЕРКАЛИ СВЕТОМ НЕБЕСА
СВЕРКАЛА ОТ БОРЬБЫ
ТВОЯ ЗАБЛУДШАЯ ДУША!
БУДУЩЕЕ ЗДЕСЬ! СОТРИ ПЕЧАЛЬ,
ОТБРОСИВ ВСЕ СОМНЕНИЯ!

ДУХ БАГРОВЫХ СЛЕЗ
ОСТАВИТ ЖАР БАГРОВЫХ ЗВЕЗД
НИ ПЛАМЯ, НИ ВЕТРА НЕ СМОГУТ СТАТЬ ПОМЕХОЙ ГРЕЗ
ТЫ ВПЕРЕД ИДИ, ОСТАВИВ СТРАХИ ПОЗАДИ КАК МЕЧТЫ...
НЕ-СОК-РУ-ШИМЫЕ!..

+++ Operation-Koppelung/RoomManager.cpp
@@ -3,7 +3,7 @@
 #include "RoomManager.h"
 #include "Menu.h"
 
-const int cache_size = 50;
+const size_t cache_size = 50;
 
 RoomManager::RoomManager(std::shared_ptr<OptionManager> optionManager) : rooms(cache_size), optionManager(optionManager) {};
